requirements.txt
sentence-transformers==2.2.0
faiss-cpu==1.7.2
openai==0.27.0
python-dotenv==0.21.0
pandas==1.5.0